requirements.txt
Flask
flask-cors
flask-jwt-extended
python-dotenv
PyPDF2
docx2txt
bcrypt
bson
pymongo